% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printline}
\alias{printline}
\title{Prints a line, without standard R formatting}
\usage{
printline(line)
}
\arguments{
\item{line}{The line to print}
}
\description{
Prints a line, without standard R formatting
}
