% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prompt_datafile}
\alias{prompt_datafile}
\title{Opens a file selection dialog box for common extensions}
\usage{
prompt_datafile()
}
\value{
The selected file, or NULL if cancelled
}
\description{
Opens a file selection dialog box for common extensions
}
