% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prompt_dir}
\alias{prompt_dir}
\title{Opens a directory selection dialog box}
\usage{
prompt_dir(default = "", caption = "Select directory")
}
\value{
The selected directory, or NULL if cancelled
}
\description{
Opens a directory selection dialog box
}
