% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prompt_file}
\alias{prompt_file}
\title{Opens a file selection dialog box}
\usage{
prompt_file(
  default = "",
  caption = "Select files",
  multi = TRUE,
  filters = NULL,
  index = 1
)
}
\value{
The selected file, or NULL if cancelled
}
\description{
Opens a file selection dialog box
}
