% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prompt_q}
\alias{prompt_q}
\title{Prompts the user with a question}
\usage{
prompt_q(q)
}
\arguments{
\item{q}{The prompt}
}
\value{
The response
}
\description{
This doesn't put a newline in, so include that if required. This does work in non-interactive mode.
}
