% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prompt_yn}
\alias{prompt_yn}
\title{Asks the user a yes or no question}
\usage{
prompt_yn(q)
}
\arguments{
\item{q}{The question}
}
\value{
TRUE on yes, FALSE on no
}
\description{
This will loop until we get a valid answer
}
