% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{random_alnum}
\alias{random_alnum}
\title{Generates a random lower-case alphanumerical string}
\usage{
random_alnum(length)
}
\arguments{
\item{length}{The length of the generated string}
}
\value{
The generated string
}
\description{
Generates a random lower-case alphanumerical string
}
