% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read.data}
\alias{read.data}
\title{Load an arbitrary input file}
\usage{
read.data(filename)
}
\arguments{
\item{filename}{Path to the input file}
}
\value{
Data parsed from filename
}
\description{
This function automatically selects the format based on the extension
}
