% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read.interactive}
\alias{read.interactive}
\title{Helps the user find the data, and then stores it for easy retrieval later}
\usage{
read.interactive()
}
\value{
The data parsed from the user
}
\description{
The selected data is stored in an Rds file in the current working directory, with a random filename
}
