% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read.interactive_inner}
\alias{read.interactive_inner}
\title{The internal logic for finding what data the user wants}
\usage{
read.interactive_inner()
}
\value{
The data parsed from the user
}
\description{
This should not be directly called, as this doesn't include the automatic saving logic
}
